fix: Folder re-structure
(Big change) Re structure of core folders, compatible with previous versions.
