+++ src/pywib/core/keyboard/__init__.py
@@ -1,0 +1,10 @@
+"""
+Keyboard functions from PyWib
+"""
+from .keyboard import (typing_speed, typing_speed_metrics, backspace_usage)
+
+__all__ = [
+    "typing_speed",
+    "typing_speed_metrics",
+    "backspace_usage",
+]
