+++ src/pywib/core/mouse/__init__.py
@@ -1,0 +1,9 @@
+"""
+Core metrics functions from PyWib
+"""
+from .mouse import click_slip, number_of_clicks
+
+__all__ = [
+    "click_slip",
+    "number_of_clicks",
+]
