+++ src/pywib/core/timing/__init__.py
@@ -1,0 +1,11 @@
+"""
+Core metrics functions from PyWib
+"""
+from .timing import execution_time, movement_time, num_pauses, pauses_metrics
+
+__all__ = [
+    "execution_time",
+    "movement_time",
+    "num_pauses",
+    "pauses_metrics",
+]
